+++ runtime/dfsan/dfsan_custom.cpp
@@ -82,13 +82,13 @@ void __taint_check_bounds(dfsan_label addr_label, uptr addr,
                           dfsan_label size_label, uint64_t size);
 
 extern "C" SANITIZER_INTERFACE_ATTRIBUTE
-void __taint_trace_cond(dfsan_label label, bool r, uint8_t flag, uint32_t cid);
+void __taint_trace_cond(dfsan_label label, bool r, uint8_t flag, uint64_t cid);
 
 extern "C" SANITIZER_INTERFACE_ATTRIBUTE
 void __taint_solve_bounds(dfsan_label ptr_label, uint64_t ptr,
                           dfsan_label index_label, int64_t index,
                           uint64_t num_elems, uint64_t elem_size,
-                          int64_t current_offset, uint32_t cid);
+                          int64_t current_offset, uint64_t cid);
 
 extern "C" SANITIZER_INTERFACE_ATTRIBUTE
 void __taint_trace_memerr(dfsan_label ptr_label, uptr ptr,
